typing

+++ poc/crypto_attacker/examples/kyber_attacker.rs
@@ -463,8 +463,8 @@ fn kyber_hacker(
 
     // instead of creating a ciphertext here, we rely on a smart process that will submit us
     // with them, we just need to measure the timing, i.e. we are implementing an oracle here
-    let oracle_repetitions: u16 = 32;
-    let mut ct_idx: u16 = 0;
+    let oracle_repetitions: usize = 32;
+    let mut ct_idx: usize = 0;
     let mut total_calls: usize = 0;
     let mut total_skipped_calls: usize = 0;
